Remove level, file, fileName, line, function, and options from DDLogMessage
The timestamp does not need to be a parameter since it is only ever
passed by the ASL log capture mechanism which we do not use and which
I believe is only for testing. This internal code path can use the ivar
accessor.

file and fileName were just duplicated data, and no one ever looks at
the fileName anyways. If you care about it, then use the tag data.

Thread name and id are also pointless to capture since I have never seen
them used for debugging. We do capture the queue name still, which can
help with debugging, and is used in places where thread ID or name were
used previously.

Create separate paths for synchronous and asynchronous messaging,
we do not need to bother with the semaphore in the synchronous path.
The messages can use the _f variants of dispatch_sync/async to avoid
having to perform a block copy; this would have made DDLog incapable
of being subclassed, but it seems the existing use of static variables
has already made that the case.

Add the Critical log level and move all log levels over such that
more levels can be added later without changing values.

Remove “Minor optimization for uniprocessor machines”. Since this case
does not apply to servers, and currently does not apply to any recent
Apple hardware either.

DDASLLogCapture will use the context for ASL filtering instead of a
bogus file name. This is faster and more sane.

Add macro variant that takes a tag object. To avoid using the DDLog
API directly in case I make more changes in the future.

The swift logging functions had to maintain a dummy entry method due
to CloudClient being required for servers…